+++ test/test_helper.rb
@@ -6,3 +6,7 @@ require "libbpf-ruby"
 require "minitest/autorun"
 
 BPF_OBJECT_PATH = File.expand_path("fixtures/test_program.bpf.o", __dir__)
+
+unless File.directory?("/sys/kernel/tracing/events")
+  system("mount", "-t", "tracefs", "tracefs", "/sys/kernel/tracing", exception: false)
+end